♻️ Add Scenario title and refactor scripts
