+++ packages/scraper/src/scrapers/CalendarScraper/scripts/2020.js
@@ -1,19 +1,6 @@
 const script = {
-  url: "https://www.chungbuk.ac.kr/site/www/sub.do?key=",
-  years: [
-    {
-      year: 2020,
-      key: "1728",
-    },
-    {
-      year: 2021,
-      key: "1804",
-    },
-    {
-      year: 2022,
-      key: "1853",
-    },
-  ],
+  url: "https://www.chungbuk.ac.kr/site/www/sub.do?key=1728",
+  yaer: 2020,
   waitCalendarSelector:
     "#contents > div.academic_calendar > ul > li > div:nth-child(2) > div:nth-child(1) > ul > li",
   getSchedules: function () {
