+++ packages/scraper/src/scrapers/DomitoryScraper/scripts/양성재.js
@@ -1,0 +1,10 @@
+const 본관 = require("./본관");
+
+const script = {
+  domitory: "양성재",
+  typeQuery: "3",
+};
+
+module.exports = { ...본관, ...script };
+
+module.exports = script;
