+++ packages/scraper/src/scrapers/DomitoryScraper/scripts/양진재.js
@@ -1,0 +1,10 @@
+const 본관 = require("./본관");
+
+const script = {
+  domitory: "양진재",
+  typeQuery: "2",
+};
+
+module.exports = { ...본관, ...script };
+
+module.exports = script;
