+++ packages/scraper/src/scrapers/NoticeScraper/index.ts
@@ -15,7 +15,7 @@ class NoticeScraper extends Scraper<NoticeScript> {
     const scripts = await this.loadScripts();
 
     scripts.forEach((script) => {
-      this.appendScenario(new Scenario(script));
+      this.appendScenario(new Scenario(script.site, script));
     });
 
     this.run();
