+++ packages/scraper/src/scrapers/Scenario.ts
@@ -1,7 +1,9 @@
 export class Scenario<T> {
+  title = "";
   jsScript?: T;
 
-  constructor(jsScript: T) {
+  constructor(title: string, jsScript: T) {
+    this.title = title;
     this.jsScript = jsScript;
   }
 }
