+++ packages/scraper/src/types/DomitoryScript.ts
@@ -2,7 +2,8 @@ export interface DomitoryScript {
   baseUrl: string;
   site: string;
   category: string;
-  domitories: { name: string; typeQuery: string }[];
+  domitory: string;
+  typeQuery: string;
   timeIndex: { morning: number; lunch: number; evening: number };
   waitMainTableSelector: string;
   getFoodList: Function;
